backend/requirements.txt
flask
python-dotenv
openai
sqlalchemy
flask-cors
